LPS-193728 Rename project to fit module name
